make builds reproducible when -seed is passed
If -seed is passed, the user wishes to explicitly provide a specific seed
for some reason. It would make sense that building the same code for
different OSes/architecures providing the same seed would result in
names getting hashed the same way.

That is not the case currently, because Build IDs are always used to
salt the hashes of names, and Build IDs differ whenever any build
arguments differ. Instead, salt hashes with the package's import path
when -seed is passed, and use Build IDs as salts otherwise. This ensures
also ensures the same name in different packages will always be hashed
differently.

Fixes #449.